+++ ssr/src/fallback.rs
@@ -27,7 +27,6 @@ pub async fn file_and_error_handler(
     }
 }
 
-#[allow(unreachable_patterns)]
 async fn get_static_file(uri: Uri, root: &str) -> Result<Response<Body>, (StatusCode, String)> {
     let req = Request::builder()
         .uri(uri.clone())
