+++ ssr/src/utils/notifications/mod.rs
@@ -1,0 +1,28 @@
+use device_id::send_principal_and_token_offchain;
+use wasm_bindgen::prelude::*;
+use wasm_bindgen_futures::JsFuture;
+
+pub mod device_id;
+
+#[wasm_bindgen(module = "/src/utils/notifications/setup-firebase-messaging.js")]
+extern "C" {
+    fn get_token(vapidKey: String) -> js_sys::Promise;
+}
+
+pub async fn get_token_for_principal(principal_id: String) {
+    let token_promise = get_token(env!("vapidKey").to_string());
+    match JsFuture::from(token_promise).await {
+        Ok(token_js) => {
+            let token: String = token_js.as_string().unwrap_or_default();
+            log::info!("sending offchain with params: {}, {}", token, principal_id);
+            send_principal_and_token_offchain(token.clone(), principal_id)
+                .await
+                .unwrap();
+            // Ok(token)
+        }
+        Err(err) => {
+            log::warn!("Failed to get token: {:?}", err);
+            // Err(())
+        }
+    }
+}
