+++ ssr/src/utils/notifications/setup-firebase-messaging.js
